Use a dedicated runner for the binary-based tests.  So far only pprof tests have been modified to demo what this would look like.
